pyproject.toml
[project]
name = "aula"
version = "0.1.0"
description = "Python API wrapper for Aula."
readme = "README.md"
authors = [
    { name = "Nick Nissen", email = "nickknissen@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.3",
    "click>=8.1.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "lxml>=5.3.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build]
packages = ["src/aula"]

[project.scripts]
aula = "aula.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/nickknissen/py-aula"
Documentation = "https://github.com/nickknissen/py-aula#readme"
Issues = "https://github.com/nickknissen/py-aula/issues"
Source = "https://github.com/nickknissen/py-aula"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "FBT", # flake8-boolean-trap
    "UP", # pyupgrade
    "YTT", # flake8-2020
    "ASYNC", # flake8-async
    "PTH", # flake8-use-pathlib
    "RUF", # ruff-specific rules
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=aula --cov-report=term-missing"

[coverage.run]
source = ["aula"]
omit = ["*/tests/*"]

[coverage.report]
show_missing = true
skip_covered = true

[coverage.html]
directory = "htmlcov"

[coverage.xml]
output = "coverage.xml"

[coverage.json]
output = "coverage.json"
